bk-001: fix code products

+++ src/variants/variants.service.ts
@@ -10,6 +10,8 @@ import { VariantEntity } from './entities/variant.entity';
 import { Repository } from 'typeorm';
 import { CreateVariantDto } from './dto/create-variants.dto';
 import { VariantItemEntity } from 'src/variant_items/entities/variant-item.entity';
+import { UserEntity } from 'src/users/entities/user.entity';
+import { UpdateVariantDto } from './dto/update-variant.dto';
 
 @Injectable()
 export class VariantsService {
@@ -20,7 +22,7 @@ export class VariantsService {
     private readonly variantItemRepository: Repository<VariantItemEntity>,
   ) {}
 
-  async create(createVariantDto: CreateVariantDto) {
+  async create(createVariantDto: CreateVariantDto, currentUser: UserEntity) {
     const variant = this.variantRepository.create(createVariantDto);
 
     //tạo variant items
@@ -30,7 +32,7 @@ export class VariantsService {
     for (let itemId of itemIds) {
       const variantItem = await this.variantItemRepository.findOne({
         where: {
-          id: itemId,
+          id: parseInt(itemId),
         },
       });
       if (variantItem) {
@@ -45,6 +47,7 @@ export class VariantsService {
     }
 
     variant.variantItems = items;
+    variant.addedBy = currentUser;
 
     return await this.variantRepository.save(variant);
   }
@@ -67,6 +70,29 @@ export class VariantsService {
     return variant;
   }
 
+  async update(
+    id: number,
+    updateVariantDto: UpdateVariantDto,
+    currentUser: UserEntity,
+  ) {
+    const variant = await this.variantRepository.findOne({
+      where: {
+        id,
+        addedBy: {
+          id: currentUser.id,
+        },
+      },
+    });
+
+    if (!variant) {
+      throw new NotFoundException('Variant not found');
+    }
+
+    Object.assign(variant, updateVariantDto);
+
+    return await this.variantRepository.save(variant);
+  }
+
   async delete(id: number) {
     const variant = await this.variantRepository.findOne({
       where: {
@@ -78,7 +104,13 @@ export class VariantsService {
     });
 
     if (!variant) {
-      throw new NotFoundException('Variant not found.');
+      throw new NotFoundException('Variant not found');
+    }
+
+    if (variant && variant?.product != null) {
+      throw new BadRequestException(
+        'Not permission to delete because of existing product!',
+      );
     }
 
     return await this.variantRepository.remove(variant);
